Titanic Dataset Logistic Regression
Here, I have used very basic implementation of logistic regression from scratch and trained it with the famous 'Titanic' dataset with the problem if which of all people can survive. I got the accuracy as 0.79, i will be visiting and updating it as i will learn more of feature scaling and EDA. Thank You.